+++ CHANGELOG.md
@@ -1,9 +1,24 @@
-## [1.0.0+3] - Wednesday 24th 2021
+## [2.0.0+5]
+September 1st 2022
 
+* Upgrade to Flutter 3.1.0
+* Use Flutter 3.3.0's Time Picker as basis
+
+
+## [1.1.0+4]
+September 1st 2022 (last Flutter 2.x.x release)
+
+* Add support for sixtieth step and an interval of 60 minutes
+* Add MIT license
+
+
+## [1.0.0+3]
+November 24th 2021
 * Null Safety release of Interval Time Picker
 * Small changes to documentation
 
 
-## [0.1.0+1] - September 2nd 2020
+## [0.1.0+1]
+September 2nd 2020
 
 * First release of Interval Time Picker